Compute coefficients *after* filter reset
The order of operations of resetting a filter type and calculating
its coefficients was upside down when a UI-initiated event reset.
This only really mattered for K35, which goes NAN on zero coefficients
because of a divide, but it was a bug for all switches.

Fix this by having the very first thing in process control to
suck the messages from the uiq and therefore reset the filter
before we do coefficients.

Closes #2